+++ docs/examples/InfiniteScrollingExample.tsx
@@ -1,12 +1,13 @@
-import { Text } from '@mantine/core';
+import { Button, Group, Text } from '@mantine/core';
 import { DataTable } from 'mantine-datatable';
-import { useEffect, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import employees from '~/data/employees.json';
 
 export default function InfiniteScrollingExample() {
-  const batchSize = 50;
+  const batchSize = 100;
   const [loading, setLoading] = useState(false);
   const [records, setRecords] = useState(employees.slice(0, batchSize));
+  const scrollViewportRef = useRef<HTMLDivElement>(null);
 
   let timeout: ReturnType<typeof setTimeout> | undefined;
 
@@ -18,6 +19,12 @@ export default function InfiniteScrollingExample() {
         setLoading(false);
       }, 1000);
     }
+  };
+
+  const reset = () => {
+    setRecords(employees.slice(0, batchSize));
+    // Make sure to scroll to top after resetting records
+    scrollViewportRef.current?.scrollTo(0, 0);
   };
 
   // Clear timeout on unmount
@@ -37,17 +44,17 @@ export default function InfiniteScrollingExample() {
         records={records}
         fetching={loading}
         onScrollToBottom={loadMoreRecords}
+        scrollViewportRef={scrollViewportRef}
       />
-      <Text mt="sm" align="center" size="sm">
-        Showing {records.length} records of {employees.length}
-      </Text>
-      <Text align="center" size="sm" color="dimmed">
-        {loading
-          ? `loading ${batchSize} more...`
-          : records.length < employees.length
-          ? 'scroll to the bottom load more'
-          : 'all records loaded'}
-      </Text>
+      <Group mt="sm" mx="xs" position="apart">
+        <Text size="sm">
+          Showing {records.length} records of {employees.length}
+          {records.length < employees.length && ', scroll to bottom to load more'}
+        </Text>
+        <Button variant="light" onClick={reset}>
+          Reset records
+        </Button>
+      </Group>
     </>
   );
 }