+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.1.2 (2023-03-16)
+
+- Add `scrollViewportRef` property to access the scroll viewport element
+- Adjust the infinite scrolling example to use the new `scrollViewportRef` property
+  
 ## 2.1.0 (2023-03-16)
 
 - Add `onScrollToTop`, `onScrollToBottom`, `onScrollToLeft` and `onScrollToRight` callbacks